+++ index.html
@@ -8,7 +8,6 @@
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
 </head>
-<script src = './js/main.js'></script>
 <body>
   <main>
 
@@ -232,5 +231,6 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
+  <script src="./js/main.js" type="module"></script>
 </body>
 </html>
